Se han hecho cambios menores a algunos apartados:
- Para el componente CComponent_Particle_Emitter:
  - Se ha definido un nuevo parámetro para el componente CComponent_Particle_Emitter:
    - Ahora, se puede concretar el número de partículas por segundo a expulsar por el emisor, evitan un posible efecto de "explosión" al inicio de la emisión.
  - Además, se ha arreglado un poco el código del componente anterior para pasar a usar VBOs (tal como se ha descrito en un comentario en el fichero de cabecera correspondiente a
dicho
componente).
- Para los game objects (CGameObject):
  - Se ha creado la función AddChildren para añadir una lista de hijos dada. Dicha lista se puede proporcionar mediante:
    - Un std::vector<CGameObject*>
    - Una lista inicializadora o sdt::initializer_list, ya sea explícito (clase anterior) o implícito mediante un array de objetos ({puntero1, puntero2...}).

+++ engine/include/engine/_engine.h
@@ -49,7 +49,7 @@ class CEngine
     void AddInstance(fboolpointer load_gameObject_function, string resource_file);
     void RemoveAllInstances();
 
-    GLfloat fps()
+    GLdouble fps()
     {
       return instances[current_instance]->fps();
     }
